Sending requested sensor measurements seems to work for now

+++ measurement.go
@@ -2,7 +2,7 @@ package main
 
 type SensorMeasurement struct {
 	SensorUUID    string
-	MeasurementId int64
+	MeasurementId int
 	Timestamp     string // Viper issue: https://github.com/spf13/viper/issues/496
 	Error         bool
 	Data          map[string]interface{}
